webapi/webapi-webspeech-w3c-tests/webspeech/blink/speech-synthesis-cancel.html

<!--
Test convert from Blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/speechsynthesis/speech-synthesis-cancel.html
-->

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Speech API Test: cancelling a speech job fires the right events</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dvcs.w3.org/hg/speech-api/raw-file/tip/speechapi.html">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var t = async_test("Test that cancelling a speech job fires the right events", { timeout: 5000});
t.step(function () {
  var u = new SpeechSynthesisUtterance("this is a test");

  u.onerror = function(event) {
    assert_false(speechSynthesis.pending);
    t.done();
  }

  speechSynthesis.speak(u);

  var u2 = new SpeechSynthesisUtterance("this is a second test");
  speechSynthesis.speak(u2);

  var u3 = new SpeechSynthesisUtterance("this is a third test");
  speechSynthesis.speak(u3);

  assert_true(speechSynthesis.pending);

  setTimeout(function () {
    speechSynthesis.cancel();
    t.done();
  }, 1);
});

</script>
